Initial extension commit
The read/write functionality for UNO API definitely works. There is a
command included as a comment in the read_from_calc() function in
extension.py that must be used to open Libreoffice with the ability to
interact with external stuff like our script.

This won't be necessary for the end user once we package everything up
into an actual extension, but to test UNO manually, it is.

Next step is to simply pass the selected data to the actual prog synth
scripts and then write it to the file.

After that we package it up and we should be good to go.